# doc/api.yaml
openapi: 3.0.0
info:
  title: WASA TEXT
  description: |
    Messagistic app
  version: 0.0.1

servers:
  - url: https://server.com/v1
  
tags:
  - name: chats operations
  - name: groups operations
  - name: login
  - name: message operations
  - name: users operations

paths:

  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserName"
        required: true
      responses:
        '200':
          description: User log-in action successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Id"
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Id"
        '400':
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"
  /users/{id}/name:
    parameters:
        - $ref: "#/components/parameters/user_id"
    put:
      tags: ["users operations"]
      operationId: setMyUserName
      security:
        - bearerAuth: []
      summary: changes the user name
      description: |
        Gets a name from query and changes the actual username to the one
        sent
      requestBody:
        description: the new name
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserName"
        required: true
      responses:
        "200":
          $ref: "#/components/responses/twohund"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"

  /users/{id}/picture:
    parameters:
        - $ref: "#/components/parameters/user_id"
    put:
      tags: ["users operations"]
      operationId: setMyPhoto
      security:
        - bearerAuth: []
      summary: changes the user picture
      description: |
        Gets a picture from query and changes the actual picture to the one
        sent
      requestBody:
        description: the user with the new picture
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          $ref: "#/components/responses/twohund"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"

  /users/{id}/search/{searchquery}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/searchquery"
    get:
      tags: ["users operations"]
      operationId: getUsers
      security:
        - bearerAuth: []
      summary: get the specified users
      description: |
        Gets the user with the name in path
      responses:
        "200": 
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersCollection"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/Fivezerozeroresp"


  /users/{id}/groups/{group_id}:
    parameters:
        - $ref: "#/components/parameters/user_id"
        - $ref: "#/components/parameters/group_id"
    delete:
      tags: ["groups operations"]
      operationId: leaveGroup
      security:
        - bearerAuth: []
      summary: leaves the group
      description: |
        Leaves the specified group
      responses:
        "200":
          $ref: "#/components/responses/twohund"
        "400":
          $ref: "#/components/responses/fourzerozero"
    put:
      tags: ["groups operations"]
      operationId: addToGroup
      security:
        - bearerAuth: []
      summary: joins the group
      description: |
        Joins the specified group
      requestBody:
        description: The user to add
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          $ref: "#/components/responses/twohund"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "505":
          $ref: "#/components/responses/fivezerozero"

  /users/{id}/groups/{group_id}/name:
    parameters:
        - $ref: "#/components/parameters/user_id"
        - $ref: "#/components/parameters/group_id"
    put:
      tags: ["groups operations"]
      operationId: setGroupName
      security:
        - bearerAuth: []
      summary: new group name
      description: |
        changes the group name based on the request body input
      requestBody:
        description: Name to change
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupName"
        required: true
      responses:
        "201":
          $ref: "#/components/responses/twozeroone"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"

  /users/{id}/groups/{group_id}/picture:
    parameters:
        - $ref: "#/components/parameters/user_id"
        - $ref: "#/components/parameters/group_id"
    put:
      tags: ["groups operations"]
      operationId: setGroupPhoto
      security:
        - bearerAuth: []
      summary: new group photo
      description: |
        changes the group photo based on the request body input
      requestBody:
        description: The chat with the new picture
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Chat"
        required: true
      responses:
        "200":
          $ref: "#/components/responses/twohund"
        "400":
          $ref: '#/components/responses/fourzerozero'
        "500":
          $ref: "#/components/responses/fivezerozero"

  /users/{id}/conversations:
    parameters:
      - $ref: "#/components/parameters/user_id"
    post:
      tags: ["chats operations"]
      operationId: newChat
      security:
        - bearerAuth: []
      summary: new chat
      description: |
        creation of a new chat, based on the input it will be a private
        chat or a group
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Chat"
        required: true
        description: Person or people to chat with
      responses:
        "201":
          description: Chat created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"
    get:
      tags: ["chats operations"]
      operationId: getMyConversations
      security:
        - bearerAuth: []
      summary: gets all the active chats
      description: |
        If there are chats it will return all the ones still active
      responses:
        "200":
          description: Chats found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCollection"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"

  /users/{id}/conversations/{chat_id}:
    parameters:
        - $ref: "#/components/parameters/user_id"
        - $ref: "#/components/parameters/chat_id"
    get:
      tags: ["chats operations"]
      operationId: getConversation
      security:
        - bearerAuth: []
      summary: gets the specified chat
      description: |
        If the chat exist it will be returned
      responses:
        "200":
          description: Chat found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"
    post:
      tags: ["message operations"]
      operationId: sendMessage
      security:
        - bearerAuth: []
      summary: sends a specified message
      description: |
        Sends the message received inside the request body
      requestBody:
        description: Message to send
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Message"
        required: true
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"

  /users/{id}/conversations/{chat_id}/message/{message_id}:
    parameters:
        - $ref: "#/components/parameters/user_id"
        - $ref: "#/components/parameters/chat_id"
        - $ref: "#/components/parameters/message_id"
    post:
      tags: ["message operations"]
      operationId: forwardMessage
      security:
        - bearerAuth: []
      summary: forwards the message
      description: |
        When on messagge path, takes chat id in input and forwards it
      requestBody:
        description: the id of the chat where to forward the message
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Id"
        required: true
      responses:
        "200":
          $ref: "#/components/responses/twohund"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"
    delete:
      tags: ["message operations"]
      operationId: deleteMessage
      security:
        - bearerAuth: []
      summary: deletes the comment
      description: |
        Deletes the message specified in the path from the chat
      responses:
        "200":
          $ref: "#/components/responses/twohund"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"

  /users/{id}/conversations/{chat_id}/message/{message_id}/comment:
    parameters:
        - $ref: "#/components/parameters/user_id"
        - $ref: "#/components/parameters/chat_id"
        - $ref: "#/components/parameters/message_id"
    put:
      tags: ["message operations"]
      operationId: commentMessage
      security:
        - bearerAuth: []
      summary: comments the message
      description: |
        When on messagge path, takes comment in input and adds it
      requestBody:
        description: the comment to write
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Text"
        required: true
      responses:
        "200":
          $ref: "#/components/responses/twohund"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"
    delete:
      tags: ["message operations"]
      operationId: uncommentMessage
      security:
        - bearerAuth: []
      summary: removes the comment
      description: |
        When on messagge path, removes the comment from it
      responses:
        "200":
          $ref: "#/components/responses/twohund"
        "400":
          $ref: "#/components/responses/fourzerozero"
        "500":
          $ref: "#/components/responses/fivezerozero"






components:
  securitySchemes:
    bearerAuth:
        type: http
        scheme: bearer
  parameters:
    user_id:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: identifier
    chat_id:
      name: chat_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: identifier
    message_id:
      name: message_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: identifier
    group_id:
      name: group_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: identifier
    searchquery:
      name: searchquery
      in: path
      required: true
      schema:
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 32
        example: Luigi
        description: Text containing the name of a user
      description: name identifier  
  schemas:
    UserName:
      type: object
      properties:
        username:
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: Maria
          description: The name of the user
      description: The name of the user
    GroupName:
      type: object
      properties:
        groupname:
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 32
          example: Calcetto
          description: The name of the group
      description: The name of the group
    Image:
      type: object
      properties:
        file_encode:
          type: string
          format: base64
          pattern: '^.*?$'
          minLength: 1
          maxLength: 9999
          description: The string in which is written the image file
      description: Profile image
    Text:
      type: object
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 80
          pattern: '^.*?$'
          description: A characters text
      description: A characters text
    Id:
      type: object
      properties:
        value:
          type: integer
          minimum: 0
          maximum: 1000
          description: The id
      description: The id
      example:
        value: 17
    User:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        usrname:
          $ref: "#/components/schemas/UserName"
        picture:
          $ref: "#/components/schemas/Image"
      description: User object
    Message:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        chat_id:
          $ref: "#/components/schemas/Id"
        sender:
          $ref: "#/components/schemas/Id"
        photo:
          $ref: "#/components/schemas/Image"
        date:
          type: string
          format: date-time
          minLength: 6
          maxLength: 10
          description: Date of send
        content:
          oneOf:
            - type: string
              pattern: '^.*?$'
              minLength: 0
              maxLength: 160
            - $ref: "#/components/schemas/Image"
          description: Content of the message
        comment:
          type: string
          minLength: 0
          maxLength: 80
          pattern: '^.*?$'
          description: Reaction to the message
      description: Message object
    Chat:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        chatname:
          type: string
          minLength: 0
          maxLength: 80
          pattern: '^.*?$'
          description: Name of the conversation
        picture:
          $ref: "#/components/schemas/Image"
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
          minItems: 1
          maxItems: 32
          description: members of the chat
        messages:
          type: array
          minItems: 0
          maxItems: 1000
          items:
            $ref: "#/components/schemas/Message"
          description: Array of messages
      required: [messages, members]
      example:
        id:
          value: 10
        chatname: calcetto
        messages:
          - Message:
              id:
                value: 0
              content: ciao
          - Message:
              id:
                value: 1
              content: tutto bene
        members:
          - User:
              id:
                value: 0
              name: Maria
          - User:
              id:
                value: 1
              name: Luigi
      description: Chat object
    ChatCollection:
      type: object
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/Chat"
          minItems: 0
          maxItems: 100
          description: Array of chats
      required: [conversations]
      description: object array of all chats
      example:
        conversations:
          - id:
              value: 0
            chatname: calcetto
            members:
              - User:
                  id:
                    value: 0
                  name: Maria
              - User:
                  id:
                    value: 1
                  name: Luigi
            messages:
              - Message:
                  id:
                    value: 0
                  content: ciao
              - Message:
                  id:
                    value: 1
                  content: tutto bene
          - id:
              value: 1
            chatname: scuola
            members:
              - User:
                  id:
                    value: 0
                  name: Maria
              - User:
                  id:
                    value: 1
                  name: Luigi
            messages:
              - Message:
                  id:
                    value: 0
                  content: ciao
              - Message:
                  id:
                    value: 1
                  content: tutto bene
    UsersCollection:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
          minItems: 0
          maxItems: 100
          description: Array of users
      required: [users]
      description: object array of all users
      example:
        users:
          - User:
              id:
                value: 0
              name: Luca
          - User:
              id:
                value: 1
              name: Giacomo
    Twozeroresp:
      type: string
      minLength: 1
      maxLength: 100
      pattern: '^.*?$'
      example: Action executed succesfully
      description: String of information
    Fourzerozeroresp:
      type: object
      properties:
        trace_four:
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 100
          example: Bad request
          description: String of information
      description: String of information
    Fourzerofourresp:
      type: object
      properties:
        trace_four:
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 100
          example: Not found
          description: String of information
      description: String of information
    fivezerozeroresp:
      type: object
      properties:
        trace_five:
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 100
          example: Server error
          description: String of information
      description: String of information
    Unauthorizedresp:
      type: object
      properties:
        trace_unauth:
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 100
          example: Access not permited
          description: String of information
      description: String of information
  responses:
    twohund:
      description: All gone correct
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Twozeroresp"
    twozeroone:
      description: Resource created
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Chat"
    fourzerozero:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Fourzerozeroresp"
    fourzerofour:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Fourzerofourresp"
    fivezerozero:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Fivezerozeroresp"
    UnauthorizedError:
      description: Access token is missing or invalid
      content: 
        application/json:
          schema:
            $ref: "#/components/schemas/Unauthorizedresp"